+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - --check-http for check TCP Tunnel after successful login
 - A separated thread to monitoring the progress of processes
 - Estimated time remaining and rate indicators
+- Targets and credentials can be loaded from multiple files
 
 ### Changed